move SessionNotFoundException to internal.exception package
