+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/core/App.java
@@ -25,12 +25,12 @@ import org.wso2.carbon.uuf.api.config.Bindings;
 import org.wso2.carbon.uuf.api.config.Configuration;
 import org.wso2.carbon.uuf.api.config.I18nResources;
 import org.wso2.carbon.uuf.api.model.MapModel;
-import org.wso2.carbon.uuf.exception.SessionNotFoundException;
 import org.wso2.carbon.uuf.exception.UUFException;
 import org.wso2.carbon.uuf.internal.exception.FragmentNotFoundException;
 import org.wso2.carbon.uuf.internal.exception.HttpErrorException;
 import org.wso2.carbon.uuf.internal.exception.PageNotFoundException;
 import org.wso2.carbon.uuf.internal.exception.PageRedirectException;
+import org.wso2.carbon.uuf.internal.exception.SessionNotFoundException;
 import org.wso2.carbon.uuf.internal.util.NameUtils;
 import org.wso2.carbon.uuf.internal.util.UriUtils;
 import org.wso2.carbon.uuf.spi.HttpRequest;
