+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/core/Page.java
@@ -19,8 +19,8 @@
 package org.wso2.carbon.uuf.core;
 
 import org.wso2.carbon.uuf.api.auth.Permission;
-import org.wso2.carbon.uuf.exception.SessionNotFoundException;
 import org.wso2.carbon.uuf.internal.debug.DebugLogger;
+import org.wso2.carbon.uuf.internal.exception.SessionNotFoundException;
 import org.wso2.carbon.uuf.internal.exception.UnauthorizedException;
 import org.wso2.carbon.uuf.internal.util.UriUtils;
 import org.wso2.carbon.uuf.spi.Renderable;
